refactor: 💡 rename error class
